+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.38.0
+- Update Count hyper prior to prevent so many underflow/overflow errors and
+    provide better fit
+- Braid should automatically convert old metadata into new metadata
+
 ## 0.37.0
 - `Engine::save` now borrows an immutable reference to the engine, thus `save`
     returns `()` and not the `Engine`.